fix: refactor package structure for better delineated models
