+++ src/eddymotion/model/__init__.py
@@ -23,17 +23,21 @@
 """Data models."""
 
 from eddymotion.model.base import (
+    AverageModel,
+    ModelFactory,
+    TrivialModel,
+)
+from eddymotion.model.dmri import (
     AverageDWModel,
     DKIModel,
     DTIModel,
     GPModel,
-    ModelFactory,
-    PETModel,
-    TrivialModel,
 )
+from eddymotion.model.pet import PETModel
 
 __all__ = (
     "ModelFactory",
+    "AverageModel",
     "AverageDWModel",
     "DKIModel",
     "DTIModel",
